rename `Heapwidget`-> `HeapWidget`
