+++ src/CMakeLists.txt
@@ -141,7 +141,7 @@ set(SOURCES
     widgets/AddressableDockWidget.cpp
     dialogs/preferences/AnalOptionsWidget.cpp
     common/DecompilerHighlighter.cpp
-        widgets/Heapwidget.cpp
+        widgets/HeapWidget.cpp
 )
 set(HEADER_FILES
     core/Cutter.h
@@ -295,7 +295,7 @@ set(HEADER_FILES
     widgets/AddressableDockWidget.h
     dialogs/preferences/AnalOptionsWidget.h
     common/DecompilerHighlighter.h
-        widgets/Heapwidget.h
+        widgets/HeapWidget.h
 )
 set(UI_FILES
     dialogs/AboutDialog.ui
@@ -362,7 +362,7 @@ set(UI_FILES
     dialogs/LayoutManager.ui
     widgets/RizinGraphWidget.ui
     dialogs/preferences/AnalOptionsWidget.ui
-        widgets/Heapwidget.ui
+        widgets/HeapWidget.ui
 )
 set(QRC_FILES
     resources.qrc
